Bugfix for `tla-gnus-article-view-patch'
Patch from Andrea Russo <rast4ndy@gmail.com>

* lisp/dvc-gnus.el (dvc-gnus-article-view-patch): Test if the article
contains a reference to an external tla patch.  Removed the
`save-window-excursion' call because it hides the diff buffer to the
user.

* lisp/tla-gnus.el (tla-gnus-article-view-patch): Added the `N'
parameter used for choosing the MIME part to pick as changeset and 
changed the documentation accordingly.
Removed the `save-excursion' call because it hides the diff buffer to
the user.