Merged various improvements from Christian Ohler <ohler+mtn@fastmail.net>
* In DVC Bookmarks Mode, C-h m now displays the key bindings available.

* Fixed wrong default for `otherdirs' in dvc-build.el so that things
no longer break when run interactively from Emacs (without going
through make).

* `dvc-add-files', `dvc-revert-files' and `dvc-remove-files' now work
as expected when the user modifies the file name at the confirmation
prompt.

* Docstring for `dvc-password-prompt-regexp' no longer incorrectly
claims that the variable is used by
eshell-watch-for-password-prompt'.

* `dvc-dvc-revision-nth-ancestor' now works for n>0.

* Added workaround for cyclic dependency between tla.el and dvc-diff.el.

* `dvc-revlist-diff' no longer calls `dvc-delta' with invalid REVISION-IDs.

* `dvc-tree-root' now returns the name of the root directory as it is,
it no longer expands all symlinks in it.  The name of the root
directory is presented to the user in various places, and presenting
it with all symlinks expanded is undesirable.  For example, if
someone's home directory is /home/foo with /home being a symlink to
/nfs/host3a14/disk0s3/home, presenting the expanded form would be
confusing.  And this is true in general: If the user wants to access a
tree through symlinks, we shouldn't be showing him a different path
(except when the disambiguation is really needed).