dvc-fileinfo: Add copied statuses.  Add `indexed' member.
Git can detect whether a file has been copied, so make provision for
this.  Also, permit the dvc-fileinfo-file struct to represent whether a
file is present in the index.  For VCS's other than git, this should
always be t.

Eventually I want to make an optional, more concise view of the status
buffer that displays the indexed field as " " if t, and "?" if nil.